requirements.txt
Flask
Flask_SQLAlchemy
Flask-Login
Werkzeug
gunicorn
pytest
Flask-Migrate
